% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConNIS.R
\name{ConNIS}
\alias{ConNIS}
\title{Calculate the probability for each gene for its biggest observed insertion free gap}
\usage{
ConNIS(
  ins.positions,
  gene.names,
  gene.starts,
  gene.stops,
  num.ins.per.gene = NULL,
  genome.length,
  weighting = 1
)
}
\arguments{
\item{ins.positions}{Numeric vector of position of observed insertions sites.}

\item{gene.names}{The names of the genes.}

\item{gene.starts}{Starting position within the genome of each gene.}

\item{gene.stops}{Ending position within the genome of each gene.}

\item{num.ins.per.gene}{Number of unique insertion sites within each gene.}

\item{genome.length}{Length of the genome.}

\item{weighting}{A weighting value for the genome-wide insertion density.}
}
\value{
The p-values for each gene to observe its biggest gap.
}
\description{
Calculate the probability for each gene for its biggest observed insertion free gap
}
\examples{
set.seed(1)
random_is <- sort(sample(1:10000, 2000))
genes <- paste("gene_", 1:30)
set.seed(2)
x <- sort(sample(1:10000, 60))
starts <- x[seq(1,60, 2)]
stops <- x[seq(2,60, 2)]
genome_length <- 10000
ConNIS(ins.positions = random_is, gene.names = genes, gene.starts = starts,
gene.stops = stops, genome.length = genome_length)
}
