% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoliData.R
\docType{data}
\name{ecoli_bw25113}
\alias{ecoli_bw25113}
\title{E. coli BW25113 data}
\format{
\subsection{\code{ecoli_bw25113}}{

A data frame with 4461 rows and 3 columns:
\describe{
\item{gene}{Name of the gene}
\item{start}{Base pair of the start of a gene}
\item{end}{Base pair of the end of a gene}
}
}
}
\source{
\url{https://biocyc.org/GCF_000750555/organism-summary}
}
\usage{
ecoli_bw25113
}
\description{
A subset from the data from E. coli BW25113 data of the BioCyc Genome
Database Collection
}
\keyword{datasets}
