% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPosData.R
\docType{data}
\name{is_pos}
\alias{is_pos}
\title{Insertion site posititions of the E.coli BW25113 strain}
\format{
\subsection{\code{is_pos}}{

A data frame with 4461 rows and 3 columns:
\describe{
\item{is_pos}{Insertion sites in base pair postions}
}
}
}
\source{
\url{https://journals.asm.org/doi/10.1128/mbio.02096-17}
}
\usage{
is_pos
}
\description{
A subset of the original from list of insertion sites of the E. coli BW25113
strain by \insertCite{Goodall2018;textual}{ConNIS}. The original lsit has
been shrunken by drawing 100000 samples.
}
\keyword{datasets}
