% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgumbel.R
\name{pgumbel}
\alias{pgumbel}
\title{Cumulative distribution function of the Gumbel distribution}
\usage{
pgumbel(k, mu, s)
}
\arguments{
\item{k}{Realsiation of the support.}

\item{mu}{The location parameter.}

\item{s}{The scale parameter.}
}
\value{
The p-values for each gene to observe its biggest gap.
}
\description{
Cumulative distribution function of the Gumbel distribution
}
\examples{
pgumbel(10,5,3)
}
