% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleResults.R
\name{subsampleResults}
\alias{subsampleResults}
\title{Function to draw subsamples of observerd IS and calculate p values}
\usage{
subsampleResults(
  method = c("Binomial", "ConNIS", "Geometric", "Tn5Gaps"),
  ins.positions,
  gene.names,
  gene.starts,
  gene.stops,
  num.ins.per.gene = NULL,
  genome.length,
  weightings = 1,
  m = 100,
  d = 0.5,
  use.parallel = FALSE,
  parallelization.type = c("mclapply", "parLapply"),
  numCores = 3,
  cluster.type = NULL,
  seed = NULL,
  rng = NULL
)
}
\arguments{
\item{method}{Method that will be applied to subsamples.}

\item{ins.positions}{Numeric vector of position of observed insertions sites.}

\item{gene.names}{The names of the genes.}

\item{gene.starts}{Starting position within the genome of each gene.}

\item{gene.stops}{Ending position within the genome of each gene.}

\item{num.ins.per.gene}{Number of unique insertion sites within each gene.}

\item{genome.length}{Length of the genome.}

\item{weightings}{A sequence of weight values that are applied to the
selected method.}

\item{m}{Number of sub samples}

\item{d}{Proportion of the original IS used as sub samples.}

\item{use.parallel}{Should the calcutltions be done in parallel?}

\item{parallelization.type}{Which method should be used for parallel
calculation. Available are \emph{mclapply} and \emph{parLapply}}

\item{numCores}{Number of cores to juse for parallel calculation. Should not
exit the number of available (logical) cores of the system.}

\item{cluster.type}{If \emph{parLapply} is used as parallelization.type, a cluster
type like \emph{PSOCK} needs tp be specified. Check your system which cluster
types are supported and see ?makeCluster of the parallel package.}

\item{seed}{Seed for subsampling. If NULL no seed is set and results might
not be reproducable.}

\item{rng}{Which random number generator (RNG) should be used. If NULL, the
default RNG of the system is used. NOTE: for parallelisation "L'Ecuyer-CMRG"
should be used}
}
\value{
A list of tibbles. Each tibble is based on one subsample and
contains the results of each genes of each tuning/weight value.
}
\description{
Function to draw subsamples of observerd IS and calculate p values
}
\examples{
set.seed(1)
random_is <- sort(sample(1:10000, 2000))
genes <- paste("gene_", 1:30)
set.seed(2)
x <- sort(sample(1:10000, 60))
starts <- x[seq(1,60, 2)]
stops <- x[seq(2,60, 2)]
genome_length <- 10000
subsampleResults(method="ConNIS",
                  ins.positions = random_is,
                  gene.names = genes,
                  gene.starts = starts,
                  gene.stops = stops,
                  genome.length = genome_length,
                  weightings = c(0.3, 0.6),
                  m = 2,
                  d = 0.5,
                  use.parallel = FALSE,
                  seed = 1)
}
