% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{Binomial}
\alias{Binomial}
\title{Calculate the probability for each gene to contain a number of observed IS
based on the binomial distribution}
\usage{
Binomial(
  ins.positions,
  gene.names,
  gene.starts,
  gene.stops,
  num.ins.per.gene = NULL,
  genome.length,
  weighting = 1
)
}
\arguments{
\item{ins.positions}{The observed unique insertion sites.}

\item{gene.names}{The names of the genes.}

\item{gene.starts}{Starting position within the genome of each gene.}

\item{gene.stops}{Ending position within the genome of each gene.}

\item{num.ins.per.gene}{Number of unique insertion sites within each gene.}

\item{genome.length}{Length of the genome.}

\item{weighting}{A weighting value for the genome-wide insertion density.}
}
\value{
The p-values for each gene to observe its biggest gap.
}
\description{
The method is an implementation of the Tn-seq analysis software (TSAS 2.0)
package \insertRef{Burger2017}{ConNIS} (see https://github.com/srimam/TSAS).
}
\examples{
set.seed(1234)
random_is <- sort(sample(1:10000, 2000))
genes <- paste("gene_", 1:30)
set.seed(5678)
x <- sort(sample(1:10000, 60))
starts <- x[seq(1,60, 2)]
stops <- x[seq(2,60, 2)]
genome_length <- 10000
Binomial(ins.positions = random_is, gene.names = genes, gene.starts = starts,
gene.stops = stops, genome.length = genome_length)
}
