% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConNIS.R
\name{ConNIS}
\alias{ConNIS}
\title{The ConNIS analysis method}
\usage{
ConNIS(
  ins.positions,
  gene.names,
  gene.starts,
  gene.stops,
  num.ins.per.gene = NULL,
  genome.length,
  weighting = 1
)
}
\arguments{
\item{ins.positions}{Numeric vector of position of observed insertions sites.}

\item{gene.names}{The names of the genes.}

\item{gene.starts}{Starting position within the genome of each gene.}

\item{gene.stops}{Ending position within the genome of each gene.}

\item{num.ins.per.gene}{Number of unique insertion sites within each gene.}

\item{genome.length}{Length of the genome.}

\item{weighting}{A weighting value for the genome-wide insertion density.}
}
\value{
The p-values for each gene to observe its biggest gap.
}
\description{
Calculates the probability for each gene of its biggest observed insertion
free gap. For a gene \eqn{j} ConNIS is defined as
\deqn{1-\sum_{i=1}^{l_j-1} \frac{\binom{b_j-i-1}{b_j - \hat{h}_j w-i}}{\binom{b_j-1}{b_j - \hat{h}_j w -1}}}
with \eqn{b_j} being the length of gene \eqn{j} and \eqn{l_j} being its
longest sequences without insertion sites. \eqn{ \hat{h}_j = b_j \cdot \theta}
are the estimated number of insertions under the assumption of
non-essentiality based on the genome-wide insertion densitiy \eqn{\theta}.
\eqn{w} we is a weight value to adjust for regions with an insertion density
below \eqn{\theta}.
}
\examples{
# generate random insertion sites
set.seed(1)
random_is <- sort(sample(1:10000, 2000))

# generate dummy gene names
genes <- paste("gene_", 1:30)

# generate radom start points and stop points of the dummy genes
set.seed(2)
x <- sort(sample(1:10000, 60))
starts <- x[seq(1,60, 2)]
stops <- x[seq(2,60, 2)]

genome_length <- 10000

ConNIS(ins.positions = random_is, gene.names = genes, gene.starts = starts,
gene.stops = stops, genome.length = genome_length)

}
