% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_seq_misses.R
\name{freq_seq_misses}
\alias{freq_seq_misses}
\title{Function to calculate the number of possibles to observe sequences of misses}
\usage{
freq_seq_misses(n, k)
}
\arguments{
\item{n}{Length of sequence.}

\item{k}{Number total misses.}
}
\value{
A list of length k with number of possibilities to observe a sequence of misses of length 1, 2, ..., k.
}
\description{
Function to calculate the number of possibles to observe sequences of misses
}
\examples{
freq_seq_misses(23,8)

}
