% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_seq_misses.R
\name{prob_seq_misses}
\alias{prob_seq_misses}
\title{Function to calculate the probabilities to observe sequences of misses}
\usage{
prob_seq_misses(n, k)
}
\arguments{
\item{n}{Length of sequence.}

\item{k}{Number total misses.}
}
\value{
A vector of length k with probabilities to observe a sequence of
misses of length 1, 2, ..., k.
}
\description{
Function to calculate the probabilities to observe sequences of misses
}
\examples{
prob_seq_misses(23,8)
}
