% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_seq_misses.R
\name{prob_seq_misses}
\alias{prob_seq_misses}
\title{Number of possible sequences of empty urns}
\usage{
prob_seq_misses(n, k)
}
\arguments{
\item{n}{Length of sequence.}

\item{k}{Number total misses.}
}
\value{
A list of length k with number of possibilities to observe a sequence of misses of length 1, 2, ..., k.
}
\description{
\code{prob_seq_misses} calculates the probability to observe \emph{s} adjacent empty
urns if \emph{n} urns are placed in a row of which \emph{k} urns are randomly
filled:
\deqn{
\frac{\binom{n-s-2}{k-s}}{\binom{n-1}{k-1}}.
}
}
\examples{
prob_seq_misses(23,8)

}
