% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tn5Gaps.R
\name{Tn5Gaps}
\alias{Tn5Gaps}
\title{TntGaps method by \insertRef{DeJesus2015}{ConNIS}
(see https://github.com/srimam/TSAS).}
\usage{
Tn5Gaps(
  ins.positions,
  gene.names,
  gene.starts,
  gene.stops,
  genome.length,
  weighting = 1
)
}
\arguments{
\item{ins.positions}{The observed unique insertion sites.}

\item{gene.names}{The names of the genes.}

\item{gene.starts}{Starting position within the genome of each gene.}

\item{gene.stops}{Ending position within the genome of each gene.}

\item{genome.length}{Length of the genome.}

\item{weighting}{A weighting value for the genome-wide insertion density.}
}
\value{
The p-values for each gene to observe its biggest gap.
}
\description{
TntGaps method by \insertRef{DeJesus2015}{ConNIS}
(see https://github.com/srimam/TSAS).
}
\examples{
set.seed(1234)
random_is <- sort(sample(1:10000, 2000))
genes <- paste("gene_", 1:30)
set.seed(5678)
x <- sort(sample(1:10000, 60))
starts <- x[seq(1,60, 2)]
stops <- x[seq(2,60, 2)]
genome_length <- 10000
Tn5Gaps(ins.positions = random_is, gene.names = genes, gene.starts = starts,
gene.stops = stops, genome.length = genome_length)
}
