% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tn5Gaps.R
\name{Tn5Gaps}
\alias{Tn5Gaps}
\title{The Tn5Gaps analysis method of TRANSIT}
\usage{
Tn5Gaps(
  ins.positions,
  gene.names,
  gene.starts,
  gene.stops,
  genome.length,
  weight = 1
)
}
\arguments{
\item{ins.positions}{Numeric vector of position of observed insertions sites.}

\item{gene.names}{The names of the genes.}

\item{gene.starts}{Starting position within the genome of each gene.}

\item{gene.stops}{Ending position within the genome of each gene.}

\item{genome.length}{Length of the genome.}

\item{weight}{A weight value for the genome-wide insertion density.}
}
\value{
The p-values for each gene to observe its biggest gap.
}
\description{
Implementation of the TntGaps method by
\insertCite{DeJesus2015;textual}{ConNIS} which uses the Gumbel distribution.
See https://transit.readthedocs.io/en/v3.2.8/method_tn5gaps.html for further
details and https://github.com/ioerger/transit for the original Python
implementation.
Here, the method is extended by allowing to set a weight value
\eqn{0 < w \leq 1} to account for genomic regions with an insertion density
below the genome-wide average.
}
\examples{
# generate random insertion sites
set.seed(1)
random_is <- sort(sample(1:10000, 2000))

# generate dummy gene names
genes <- paste("gene_", 1:30)

# generate radom start points and stop points of the dummy genes
set.seed(2)
x <- sort(sample(1:10000, 60))
starts <- x[seq(1,60, 2)]
stops <- x[seq(2,60, 2)]

genome_length <- 10000

Tn5Gaps(ins.positions = random_is, gene.names = genes, gene.starts = starts,
gene.stops = stops, genome.length = genome_length)

}
\references{
\insertRef{DeJesus2015}{ConNIS}
}
